COM-3273: use composable
