+++ src/core/components/courses/ProfileOrganization.vue
@@ -33,7 +33,7 @@
     <ni-slot-container :can-edit="canEditSlots" :loading="courseLoading" @refresh="refreshCourse" :is-admin="isAdmin"
       @update="updateCourse('estimatedStartDate')" />
     <ni-company-table v-if="isCourseInter && isVendorInterface" :companies="course.companies"
-      @refresh="refreshCourse" :can-edit="canEditCompanies" :loading="courseLoading" />
+      @refresh="refreshCourse" :can-edit="isAdmin" :loading="courseLoading" />
     <ni-trainee-table :can-edit="canEditTrainees" :loading="courseLoading" @refresh="refreshCourse"
       @update="updateCourse('maxTrainees')" :validations="v$.course" />
     <q-page-sticky expand position="right">
@@ -258,8 +258,6 @@ export default {
     const canEditSlots = computed(() => !(isClientInterface && isCourseInter.value));
 
     const canEditTrainees = computed(() => isIntraCourse.value || (!isClientInterface && !isTrainer.value));
-
-    const canEditCompanies = computed(() => isCourseInter.value && !isClientInterface && !isTrainer.value);
 
     const isFinished = computed(() => {
       const slotsToCome = course.value.slots.filter(slot => CompaniDate().isBefore(slot.endDate));
@@ -740,7 +738,6 @@ export default {
       isAdmin,
       canEditSlots,
       canEditTrainees,
-      canEditCompanies,
       filteredMessageTypeOptions,
       missingTraineesPhone,
       smsMissingInfo,
